initial trace skeleton

+++ src/runtime_src/xdp/profile/plugin/aie_trace/aie_trace_impl.h
@@ -46,7 +46,7 @@ namespace xdp {
 
     virtual void updateDevice() = 0;
     virtual void freeResources() = 0;
-    virtual void pollTimers(uint32_t index, void* handle) = 0;
+    virtual void pollTimers(uint64_t index, void* handle) = 0;
     virtual uint64_t checkTraceBufSize(uint64_t size) = 0;
     /*
      * If trace module is running, it might buffer partial trace.
